+++ packages/account/SimpleAccount/index.ts
@@ -4,6 +4,7 @@ import number from "~packages/util/number"
 import type { BigNumberish, HexString } from "~typing"
 
 import type { Account, Call } from "../index"
+import { SimpleAccountFactory } from "./SimpleAccountFactory"
 
 export class SimpleAccount implements Account {
   /**
@@ -116,54 +117,5 @@ export class SimpleAccount implements Account {
       throw new Error("No factory")
     }
     return this.factory
-  }
-}
-
-class SimpleAccountFactory {
-  private node: ethers.JsonRpcProvider
-
-  private factory: ethers.Contract
-  private owner: HexString
-  private salt: BigNumberish
-
-  public constructor(opts: {
-    address: string
-    owner: HexString
-    salt: BigNumberish
-    nodeRpcUrl: string
-  }) {
-    this.node = new ethers.JsonRpcProvider(opts.nodeRpcUrl)
-    this.factory = new ethers.Contract(
-      opts.address,
-      [
-        "function getAddress(address owner, uint256 salt) view returns (address)",
-        "function createAccount(address owner,uint256 salt)"
-      ],
-      this.node
-    )
-    this.owner = opts.owner
-    this.salt = opts.salt
-  }
-
-  public async getAddress() {
-    return ethers.zeroPadValue(
-      ethers.stripZerosLeft(
-        // The name of `getAddress` conflicts with the function on ethers.Contract.
-        // So we build call data from interface and directly send through node rpc provider.
-        await this.node.call(
-          await this.factory
-            .getFunction("getAddress")
-            .populateTransaction(this.owner, this.salt)
-        )
-      ),
-      20
-    )
-  }
-
-  public async getInitCode() {
-    const { data } = await this.factory
-      .getFunction("createAccount")
-      .populateTransaction(this.owner, this.salt)
-    return ethers.concat([await this.factory.getAddress(), data])
   }
 }
